% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc.R
\name{auc}
\alias{auc}
\title{auc}
\usage{
auc(object, multimodal = FALSE, prob = 0.95, adjust = 1, summary_only = FALSE)
}
\arguments{
\item{object}{An object of class \code{curvish.curve}}

\item{multimodal}{if the expected posterior might be multimodal, set this to
\code{TRUE} so that the summary is a true, possibly discontinuous, highest
posterior density interval.}

\item{prob}{the proportion of mass within the HPDI}

\item{adjust}{passed to \code{\link[stats]{density}} to adjust bandwidth.}

\item{summary_only}{set this to \code{FALSE} if you want all samples
returned.}
}
\value{

}
\description{
Find the area under the curve of the derivative. This function uses the
trapezoidal method.
}
