% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curvish-package.R
\docType{package}
\name{curvish-package}
\alias{curvish}
\alias{curvish-package}
\title{curvish: Uses Bayesian GAMMs to Regions Where Slopes Differ}
\description{
Uses Bayesian generalized additive mixed models to estimate 
  curves through data and the uses the posterior derivative to compare points
  along those curves in order to determine regions of difference from steepest
  slope.
}
\author{
\strong{Maintainer}: John Flournoy \email{john_flournoy@g.harvard.edu}

}
\keyword{internal}
