% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivatives.R
\name{derivatives}
\alias{derivatives}
\title{derivatives}
\usage{
derivatives(object, term, n = 200, eps = 1e-07, summary_only = TRUE)
}
\arguments{
\item{object}{a \code{brmsfit} object}

\item{term}{character; the smooth term for which the derivative is required.}

\item{n}{numeric; the number of points to evaluate the derivative at.}

\item{eps}{numeric; the finite difference.}
}
\value{

}
\description{
Computes the derivative of a smooth term from the posterior of the smooth.
}
