% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_X_at_slope_point.R
\name{get_X_at_slope_point}
\alias{get_X_at_slope_point}
\title{Title}
\usage{
get_X_at_slope_point(
  object,
  term,
  n,
  eps,
  find = c("min", "max", "value"),
  value = NULL,
  multimodal = FALSE,
  prob = 0.95,
  summary_only = TRUE
)
}
\arguments{
\item{object}{a \code{brmsfit} object}

\item{term}{character; the smooth term for which the derivative is required.}

\item{n}{numeric; the number of points to evaluate the derivative at.}

\item{eps}{numeric; the finite difference.}

\item{find}{the spot along the derivative to find; can be 'min', 'max', or 'value', in which case the \code{value} parameter must be set.}

\item{value}{if \code{find} is set to 'value', this must be set to a specific value, like 0.}

\item{multimodal}{if the expected posterior might be multimodal, set this to \code{TRUE} so that the summary is a true, possibly discontinuous, highest posterior density interval.}

\item{prob}{the proportion of mass within the HPDI}

\item{summary_only}{set this to \code{FALSE} if you want all samples returned.}
}
\value{
either a summary of the posterior density, or a list containing the summary and the posterior samples.
}
\description{
Find the posterior distribution of the value of a predictor variable at a
particular point along the derivative. For instance, one might be interested
in knowing the age value at which a slope is steepest, or closest to zero.
}
