% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{plot.curvish.param}
\alias{plot.curvish.param}
\title{plot.curvish.param}
\usage{
\method{plot}{curvish.param}(x, robust = FALSE, range = NULL, mode = TRUE, ...)
}
\arguments{
\item{x}{An object of class \code{curvish.param}.}

\item{robust}{If the type of summary is multimodal and this is \code{TRUE},
this sets the parameter of central tendency to the median.}

\item{range}{Two-element numeric vector specifying limits on the range of the
posterior to consider for the plot.}

\item{mode}{If the type of summary is multimodal, default is to use mode as
the central tendency. If this is \code{FALSE}, will use the median of the
posterior within the relevant portion of the HPDI.}

\item{...}{Will be passed to the \code{density} function.}
}
\value{

}
\description{
plot.curvish.param
}
