% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_X_at_slope_point.R
\name{posterior_x_at_fy}
\alias{posterior_x_at_fy}
\alias{posterior_x_at_miny}
\alias{posterior_x_at_maxy}
\alias{posterior_x_at_yequalto}
\title{posterior_x_at_fy}
\usage{
posterior_x_at_fy(
  object,
  term,
  n,
  eps,
  find = c("min", "max", "value"),
  value = NULL,
  multimodal = TRUE,
  adjust = 1,
  prob = 0.95,
  summary_only = FALSE
)

posterior_x_at_miny(...)

posterior_x_at_maxy(...)

posterior_x_at_yequalto(...)
}
\arguments{
\item{object}{a \code{brmsfit} object}

\item{term}{character; the smooth term for which the derivative is required.}

\item{n}{numeric; the number of points to evaluate the derivative at.}

\item{eps}{numeric; the finite difference.}

\item{find}{the spot along the derivative to find; can be 'min', 'max', or
'value', in which case the \code{value} parameter must be set.}

\item{value}{if \code{find} is set to 'value', this must be set to a specific
value, like 0.}

\item{multimodal}{if the expected posterior might be multimodal, set this to
\code{TRUE} so that the summary is a true, possibly discontinuous, highest
posterior density interval.}

\item{adjust}{passed to \code{\link[stats]{density}} to adjust bandwidth.}

\item{prob}{the proportion of mass within the HPDI}

\item{summary_only}{set this to \code{FALSE} if you want all samples
returned.}
}
\value{
An object of class \code{curvish.param} that contains the summary
  from \code{\link[HDInterval]{hdi}}, and the posterior samples. If
  \code{summary_only} is \code{TRUE}, the only the output from
  \code{\link[HDInterval]{hdi}} is returned.
}
\description{
Find the posterior distribution of the value of a predictor variable at a
particular point along the derivative. For instance, one might be interested
in knowing the age value at which a slope is steepest, or closest to zero.
}
\details{
\code{posterior_x_at_miny}, \code{posterior_x_at_maxy}, and
\code{posterior_x_at_yequalto} are aliases of \code{posterior_x_at_fy} where
\code{find} is set to 'min', 'max', or 'value', respectively and are provided
in case they make for more readable code.
}
