% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivatives.R
\name{derivatives}
\alias{derivatives}
\title{derivatives}
\usage{
derivatives(
  object,
  term,
  n = 200,
  eps = 1e-05,
  prob = 0.95,
  prob_outer = 0.99,
  order = 1,
  deriv_posterior = TRUE
)
}
\arguments{
\item{object}{a \code{brmsfit} object}

\item{term}{character; the smooth term for which the derivative is required.}

\item{n}{numeric; the number of points to evaluate the derivative at.}

\item{eps}{numeric; the finite difference.}

\item{prob}{numeric; the density of the posterior summary.}

\item{prob_outer}{numeric; the outer density of the posterior summary.}

\item{order}{integer; the order of the derivative. Only first (\code{1}) and
second (\code{2}) are supported.}

\item{deriv_posterior}{return the posterior samples for the derivative as
part of the object?}
}
\value{
a \code{curvish} object containing summaries of the posterior
  derivative, posterior smooths.
}
\description{
Computes the derivative of a smooth term from the posterior of the smooth.
}
